backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.0",
    "google-generativeai>=0.8.5",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
